@a-ghorbani 2025-03-20
Hey, is this implementation expected to give a boost on mobile devices as well (which afaik usually have a 128-bit SVE vector), or is this meant for server-grade CPUs (with larger SVE widths)?I tested it on my Pixel 9 (which has SVE and SVE2 at a 128-bit) and couldn't see any performance improvement. so wanted to check if I am missing something or it is expected.