@fj-y-saito 2025-03-24
This imprementation will improve performance if the processor meets the following conditions:
## SIMD width is 128
- Number of NEON pipelines < Number of SVE pipelines

## SIMD width is 256 or more
- Number of NEON pipelines < Number of SVE pipelines x 2