@BMalaca 2025-11-10
@fj-y-saito what compiler did you use for these tests? I'm trying to replicate the tg128 result in a FX700 but your results are 50% faster in the tg128. Also, if you could share your cmake options that'd be great. Thanks!