@fj-y-saito 2025-11-13
I used gcc 11.3.0 for the tests.
I didn’t specify any special options for CMake (cmake -D...), just the default build.

At the moment, I'm not sure why the performance differs in this specific test.
If you could share your compiler version and the CMake options you used,
that would help me compare our setups.